Creates separate file for ImageLabelData
