+++ CustomDropDown/Model/ImageLabelData.swift
@@ -1,0 +1,19 @@
+//
+//  ImageLabelData.swift
+//  CustomDropDown
+//
+//  Created by Jacqueline Alves on 22/10/20.
+//  Copyright © 2020 Amirthy Tejeshwar. All rights reserved.
+//
+
+import UIKit
+
+public struct ImageLabelData {
+    public var title: String?
+    public var image: UIImage?
+    
+    public init(title: String?, image: UIImage?) {
+        self.title = title
+        self.image = image
+    }
+}
